requirements.txt
pymongo[srv]
from-root
scikit-learn
numpy
pandas
python-dotenv
black
dill
imblearn
xgboost
scipy
evidently
PyYAML
-e .